+++ libpod/networking_unsupported.go
@@ -2,10 +2,6 @@
 
 package libpod
 
-import (
-	"github.com/containers/libpod/pkg/inspect"
-)
-
 func (r *Runtime) setupRootlessNetNS(ctr *Container) (err error) {
 	return ErrNotImplemented
 }
@@ -22,6 +18,6 @@ func (r *Runtime) createNetNS(ctr *Container) (err error) {
 	return ErrNotImplemented
 }
 
-func (c *Container) getContainerNetworkInfo(data *inspect.ContainerInspectData) *inspect.ContainerInspectData {
+func (c *Container) getContainerNetworkInfo(data *InspectContainerData) *InspectContainerData {
 	return nil
 }
